+++ examples/run_websocket_async.py
@@ -31,13 +31,14 @@ async def _message_handler(event, message: qqbot.Message):
     :param event: 事件类型
     :param message: 事件对象（如监听消息是Message对象）
     """
-    msg_api = qqbot.MessageAPI(t_token, False)
+    msg_api = qqbot.AsyncMessageAPI(t_token, False)
     # 打印返回信息
     qqbot.logger.info("event %s" % event + ",receive message %s" % message.content)
-    # 构造消息发送请求数据对象
-    send = qqbot.MessageSendRequest("<@%s>谢谢你，加油 " % message.author.id, message.id)
-    # 通过api发送回复消息
-    msg_api.post_message(message.channel_id, send)
+    for i in range(100):
+        # 构造消息发送请求数据对象
+        send = qqbot.MessageSendRequest("<@%s>谢谢你，加油 " % message.author.id, message.id)
+        # 通过api发送回复消息
+        await msg_api.post_message(message.channel_id, send)
 
 
 if __name__ == "__main__":
