+++ qqbot/core/network/ws/ws_handler.py
