+++ qqbot/core/network/ws/ws_session.py
